+++ opentelemetry-instrumentation/src/opentelemetry/instrumentation/instrumentor.py
@@ -21,6 +21,9 @@ from abc import ABC, abstractmethod
 from logging import getLogger
 from typing import Collection, Optional
 
+from opentelemetry.instrumentation._semconv import (
+    _OpenTelemetrySemanticConventionStability,
+)
 from opentelemetry.instrumentation.dependencies import (
     DependencyConflict,
     get_dependency_conflicts,
@@ -104,6 +107,9 @@ class BaseInstrumentor(ABC):
             if conflict:
                 _LOG.error(conflict)
                 return None
+
+        # initialize semantic conventions opt-in if needed
+        _OpenTelemetrySemanticConventionStability._initialize()
 
         result = self._instrument(  # pylint: disable=assignment-from-no-return
             **kwargs